requirements.txt
Pillow
enquiries